+++ lib/auth/authentication.dart
@@ -3,5 +3,5 @@ part of dart_openapi;
 abstract class Authentication {
   /// Apply authentication settings to header and query params.
   void applyToParams(
-      List<QueryParam> queryParams, Map<String, String> headerParams);
+      List<QueryParam> queryParams, Map<String, dynamic> headerParams);
 }