+++ lib/auth/http_basic_auth.dart
@@ -6,7 +6,7 @@ class HttpBasicAuth implements Authentication {
 
   @override
   void applyToParams(
-      List<QueryParam> queryParams, Map<String, String> headerParams) {
+      List<QueryParam> queryParams, Map<String, dynamic> headerParams) {
     String str = (_username == null ? "" : _username) +
         ":" +
         (_password == null ? "" : _password);