+++ lib/auth/authentication.dart
@@ -1,4 +1,5 @@
 part of dart_openapi;
+// @dart=2.9
 
 abstract class Authentication {
   /// Apply authentication settings to header and query params.