+++ lib/openapi.dart
@@ -1,24 +1,18 @@
 library dart_openapi;
+// @dart=2.9
 
 import 'dart:async';
 import 'dart:convert';
+import 'dart:core';
 
 import 'package:dio/dio.dart';
 
-part 'auth/api_key_auth.dart';
-
-part 'auth/authentication.dart';
-
-part 'auth/http_basic_auth.dart';
-
-part 'auth/oauth.dart';
-
 part 'api_client.dart';
-
-part 'api_helper.dart';
-
-part 'api_exception.dart';
-
 part 'api_delegates.dart';
-
 part 'api_dio_client_delegate.dart';
+part 'api_exception.dart';
+part 'api_helper.dart';
+part 'auth/api_key_auth.dart';
+part 'auth/authentication.dart';
+part 'auth/http_basic_auth.dart';
+part 'auth/oauth.dart';